[144] Fix Add existing elements tool recursive on PartUsage
Bug: https://github.com/eclipse-syson/syson/issues/144
